requirements.txt
anthropic>=0.39.0
openai>=1.50.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pypdf>=3.0.0
python-docx>=1.0.0
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.0.0
